Moved Campaign Options into own package
