+++ MekHQ/src/mekhq/campaign/personnel/procreation/DisabledRandomProcreation.java
@@ -24,14 +24,19 @@
  *
  * Catalyst Game Labs and the Catalyst Game Labs logo are trademarks of
  * InMediaRes Productions, LLC.
+ *
+ * MechWarrior Copyright Microsoft Corporation. MekHQ was created under
+ * Microsoft's "Game Content Usage Rules"
+ * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
+ * affiliated with Microsoft.
  */
 package mekhq.campaign.personnel.procreation;
 
-import mekhq.campaign.CampaignOptions;
+import java.time.LocalDate;
+
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.enums.RandomProcreationMethod;
-
-import java.time.LocalDate;
 
 public class DisabledRandomProcreation extends AbstractProcreation {
     //region Constructors
