+++ MekHQ/unittests/mekhq/campaign/market/ContractMarketAtBGenerationTests.java
@@ -53,10 +53,10 @@ import java.util.UUID;
 import java.util.Vector;
 
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.CurrentLocation;
 import mekhq.campaign.Hangar;
 import mekhq.campaign.JumpPath;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.finances.Accountant;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.force.Force;
