+++ MekHQ/unittests/mekhq/campaign/parts/equipment/EquipmentPartTest.java
@@ -24,28 +24,23 @@
  *
  * Catalyst Game Labs and the Catalyst Game Labs logo are trademarks of
  * InMediaRes Productions, LLC.
+ *
+ * MechWarrior Copyright Microsoft Corporation. MekHQ was created under
+ * Microsoft's "Game Content Usage Rules"
+ * <https://www.xbox.com/en-US/developers/rules> and it is not endorsed by or
+ * affiliated with Microsoft.
  */
 package mekhq.campaign.parts.equipment;
 
-import megamek.Version;
-import megamek.common.*;
-import megamek.common.equipment.WeaponMounted;
-import megamek.common.weapons.bayweapons.BayWeapon;
-import mekhq.utilities.MHQXMLUtility;
-import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
-import mekhq.campaign.Quartermaster;
-import mekhq.campaign.Warehouse;
-import mekhq.campaign.parts.Part;
-import mekhq.campaign.unit.Unit;
-import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.xml.sax.SAXException;
+import static mekhq.campaign.parts.equipment.EquipmentUtilities.getEquipmentType;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.*;
 
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.ParserConfigurationException;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -53,10 +48,25 @@ import java.io.StringWriter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Vector;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.ParserConfigurationException;
 
-import static mekhq.campaign.parts.equipment.EquipmentUtilities.getEquipmentType;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import megamek.Version;
+import megamek.common.*;
+import megamek.common.equipment.WeaponMounted;
+import megamek.common.weapons.bayweapons.BayWeapon;
+import mekhq.campaign.Campaign;
+import mekhq.campaign.Quartermaster;
+import mekhq.campaign.Warehouse;
+import mekhq.campaign.campaignOptions.CampaignOptions;
+import mekhq.campaign.parts.Part;
+import mekhq.campaign.unit.Unit;
+import mekhq.utilities.MHQXMLUtility;
+import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.xml.sax.SAXException;
 
 public class EquipmentPartTest {
     @Test
